server_py/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
requests==2.32.3
python-multipart==0.0.9
oss2==2.18.1
SQLAlchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
